# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Industry Research Pod API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: companies
    description: Company Intelligence Cards
  - name: shei
    description: SHEI Problem Hypothesis Cards
  - name: signals
    description: Signal tracking
  - name: benchmarks
    description: KPI Benchmarking Snapshots
  - name: playbooks
    description: Functional Playbooks
  - name: dashboard
    description: Dashboard summaries

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies
      parameters:
        - name: tier
          in: query
          schema:
            type: string
            enum: [1, 2]
        - name: geography
          in: query
          schema:
            type: string
            enum: [INDIA, GLOBAL]
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /shei-cards:
    get:
      operationId: listSheiCards
      tags: [shei]
      summary: List SHEI hypothesis cards
      parameters:
        - name: functionTag
          in: query
          schema:
            type: string
        - name: geography
          in: query
          schema:
            type: string
        - name: urgency
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of SHEI cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SheiCard"
    post:
      operationId: createSheiCard
      tags: [shei]
      summary: Create a SHEI card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSheiCardBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheiCard"

  /shei-cards/{id}:
    get:
      operationId: getSheiCard
      tags: [shei]
      summary: Get SHEI card by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SHEI card detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheiCard"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateSheiCard
      tags: [shei]
      summary: Update a SHEI card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSheiCardBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheiCard"
    delete:
      operationId: deleteSheiCard
      tags: [shei]
      summary: Delete a SHEI card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /signals:
    get:
      operationId: listSignals
      tags: [signals]
      summary: List signals
      parameters:
        - name: companyId
          in: query
          schema:
            type: integer
        - name: strength
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
    post:
      operationId: createSignal
      tags: [signals]
      summary: Create a signal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSignalBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"

  /signals/{id}:
    delete:
      operationId: deleteSignal
      tags: [signals]
      summary: Delete a signal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /benchmarks:
    get:
      operationId: listBenchmarks
      tags: [benchmarks]
      summary: List KPI benchmarks
      parameters:
        - name: functionTag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of benchmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Benchmark"
    post:
      operationId: createBenchmark
      tags: [benchmarks]
      summary: Create a benchmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBenchmarkBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Benchmark"

  /benchmarks/{id}:
    put:
      operationId: updateBenchmark
      tags: [benchmarks]
      summary: Update a benchmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBenchmarkBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Benchmark"
    delete:
      operationId: deleteBenchmark
      tags: [benchmarks]
      summary: Delete a benchmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /playbooks:
    get:
      operationId: listPlaybooks
      tags: [playbooks]
      summary: List playbook sections
      parameters:
        - name: function
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of playbook sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaybookSection"
    post:
      operationId: createPlaybookSection
      tags: [playbooks]
      summary: Create a playbook section
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlaybookSectionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybookSection"

  /playbooks/{id}:
    put:
      operationId: updatePlaybookSection
      tags: [playbooks]
      summary: Update a playbook section
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlaybookSectionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybookSection"
    delete:
      operationId: deletePlaybookSection
      tags: [playbooks]
      summary: Delete a playbook section
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Dashboard overview summary
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/recent-signals:
    get:
      operationId: getRecentSignals
      tags: [dashboard]
      summary: Recent high-signal intelligence items
      responses:
        "200":
          description: Recent signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"

  /dashboard/signal-breakdown:
    get:
      operationId: getSignalBreakdown
      tags: [dashboard]
      summary: Signal counts by category
      responses:
        "200":
          description: Signal breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SignalCategoryCount"

  /dashboard/shei-by-urgency:
    get:
      operationId: getSheiByUrgency
      tags: [dashboard]
      summary: SHEI cards grouped by urgency
      responses:
        "200":
          description: SHEI urgency breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SheiUrgencyCount"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        fullName:
          type: string
        tier:
          type: integer
          enum: [1, 2]
        geography:
          type: string
          enum: [INDIA, GLOBAL]
        exchange:
          type: string
        irPage:
          type: string
        earningsCadence:
          type: string
        revenue:
          type: string
        revenueGrowth:
          type: string
        ebitdaMargin:
          type: string
        marketCap:
          type: string
        categories:
          type: string
        strategicPriorities:
          type: string
        scIntelligence:
          type: string
        techIntelligence:
          type: string
        openProblems:
          type: string
        quickTake:
          type: string
        confidence:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        version:
          type: string
        nextRefresh:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, fullName, tier, geography, confidence, version, createdAt, updatedAt]

    CreateCompanyBody:
      type: object
      properties:
        name:
          type: string
        fullName:
          type: string
        tier:
          type: integer
          enum: [1, 2]
        geography:
          type: string
          enum: [INDIA, GLOBAL]
        exchange:
          type: string
        irPage:
          type: string
        earningsCadence:
          type: string
        revenue:
          type: string
        revenueGrowth:
          type: string
        ebitdaMargin:
          type: string
        marketCap:
          type: string
        categories:
          type: string
        strategicPriorities:
          type: string
        scIntelligence:
          type: string
        techIntelligence:
          type: string
        openProblems:
          type: string
        quickTake:
          type: string
        confidence:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        version:
          type: string
        nextRefresh:
          type: string
      required: [name, fullName, tier, geography]

    SheiCard:
      type: object
      properties:
        id:
          type: integer
        cardId:
          type: string
        title:
          type: string
        functionTag:
          type: string
          enum: [SUPPLY_CHAIN, IT_TECHNOLOGY, SALES_DISTRIBUTION, FINANCIAL, CATEGORY]
        geographyTag:
          type: string
          enum: [INDIA, GLOBAL, BOTH]
        urgency:
          type: string
          enum: [IMMEDIATE, MEDIUM_TERM, STRUCTURAL]
        signal:
          type: string
        hypothesis:
          type: string
        evidence:
          type: string
        clientImplication:
          type: string
        thoucentriqAngle:
          type: string
        pitchAnchor:
          type: string
        provocQuestion:
          type: string
        povParagraph:
          type: string
        version:
          type: string
        nextReview:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, cardId, title, functionTag, geographyTag, urgency, signal, hypothesis, evidence, clientImplication, thoucentriqAngle, version, createdAt, updatedAt]

    CreateSheiCardBody:
      type: object
      properties:
        cardId:
          type: string
        title:
          type: string
        functionTag:
          type: string
          enum: [SUPPLY_CHAIN, IT_TECHNOLOGY, SALES_DISTRIBUTION, FINANCIAL, CATEGORY]
        geographyTag:
          type: string
          enum: [INDIA, GLOBAL, BOTH]
        urgency:
          type: string
          enum: [IMMEDIATE, MEDIUM_TERM, STRUCTURAL]
        signal:
          type: string
        hypothesis:
          type: string
        evidence:
          type: string
        clientImplication:
          type: string
        thoucentriqAngle:
          type: string
        pitchAnchor:
          type: string
        provocQuestion:
          type: string
        povParagraph:
          type: string
        version:
          type: string
        nextReview:
          type: string
      required: [cardId, title, functionTag, geographyTag, urgency, signal, hypothesis, evidence, clientImplication, thoucentriqAngle]

    Signal:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        companyName:
          type: string
        category:
          type: string
          enum: [EARNINGS, STRATEGY, SUPPLY_CHAIN, TECHNOLOGY, LEADERSHIP, REGULATORY, DISRUPTION, MACRO]
        summary:
          type: string
        strength:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        action:
          type: string
          enum: [UPDATE_CARD, SHEI_CANDIDATE, DISRUPTION_BRIEF, MONITOR, NO_ACTION]
        source:
          type: string
        createdAt:
          type: string
      required: [id, category, summary, strength, action, createdAt]

    CreateSignalBody:
      type: object
      properties:
        companyId:
          type: integer
        companyName:
          type: string
        category:
          type: string
          enum: [EARNINGS, STRATEGY, SUPPLY_CHAIN, TECHNOLOGY, LEADERSHIP, REGULATORY, DISRUPTION, MACRO]
        summary:
          type: string
        strength:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        action:
          type: string
          enum: [UPDATE_CARD, SHEI_CANDIDATE, DISRUPTION_BRIEF, MONITOR, NO_ACTION]
        source:
          type: string
      required: [category, summary, strength, action]

    Benchmark:
      type: object
      properties:
        id:
          type: integer
        kpiName:
          type: string
        definition:
          type: string
        functionTag:
          type: string
        bestInClass:
          type: string
        industryMedian:
          type: string
        laggard:
          type: string
        indiaContext:
          type: string
        sheiAnnotation:
          type: string
        version:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, kpiName, definition, functionTag, bestInClass, industryMedian, laggard, version, createdAt, updatedAt]

    CreateBenchmarkBody:
      type: object
      properties:
        kpiName:
          type: string
        definition:
          type: string
        functionTag:
          type: string
        bestInClass:
          type: string
        industryMedian:
          type: string
        laggard:
          type: string
        indiaContext:
          type: string
        sheiAnnotation:
          type: string
        version:
          type: string
      required: [kpiName, definition, functionTag, bestInClass, industryMedian, laggard]

    PlaybookSection:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        functionTag:
          type: string
        sectionNumber:
          type: integer
        whyItMatters:
          type: string
        industryLandscape:
          type: string
        failureModes:
          type: string
        whatGoodLooksLike:
          type: string
        technologyEnablers:
          type: string
        consultingEntryPoints:
          type: string
        status:
          type: string
          enum: [COMPLETE, STARTER, IN_PROGRESS]
        version:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, functionTag, sectionNumber, whyItMatters, status, version, createdAt, updatedAt]

    CreatePlaybookSectionBody:
      type: object
      properties:
        title:
          type: string
        functionTag:
          type: string
        sectionNumber:
          type: integer
        whyItMatters:
          type: string
        industryLandscape:
          type: string
        failureModes:
          type: string
        whatGoodLooksLike:
          type: string
        technologyEnablers:
          type: string
        consultingEntryPoints:
          type: string
        status:
          type: string
          enum: [COMPLETE, STARTER, IN_PROGRESS]
        version:
          type: string
      required: [title, functionTag, sectionNumber, whyItMatters, status]

    DashboardSummary:
      type: object
      properties:
        totalCompanies:
          type: integer
        tier1Companies:
          type: integer
        tier2Companies:
          type: integer
        totalSheiCards:
          type: integer
        immediateSheiCards:
          type: integer
        totalSignals:
          type: integer
        highSignals:
          type: integer
        totalBenchmarks:
          type: integer
        totalPlaybookSections:
          type: integer
      required: [totalCompanies, tier1Companies, tier2Companies, totalSheiCards, immediateSheiCards, totalSignals, highSignals, totalBenchmarks, totalPlaybookSections]

    SignalCategoryCount:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
      required: [category, count]

    SheiUrgencyCount:
      type: object
      properties:
        urgency:
          type: string
        count:
          type: integer
      required: [urgency, count]