# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Industry Research Pod API
 * OpenAPI spec version: 0.1.0
 */
import type { CompanyConfidence } from "./companyConfidence";
import type { CompanyGeography } from "./companyGeography";
import type { CompanyTier } from "./companyTier";

export interface Company {
  id: number;
  name: string;
  fullName: string;
  tier: CompanyTier;
  geography: CompanyGeography;
  exchange?: string;
  irPage?: string;
  earningsCadence?: string;
  revenue?: string;
  revenueGrowth?: string;
  ebitdaMargin?: string;
  marketCap?: string;
  categories?: string;
  strategicPriorities?: string;
  scIntelligence?: string;
  techIntelligence?: string;
  openProblems?: string;
  quickTake?: string;
  confidence: CompanyConfidence;
  version: string;
  nextRefresh?: string;
  createdAt: string;
  updatedAt: string;
}