# lib/api-zod/src/generated/types/createCompanyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Industry Research Pod API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCompanyBodyConfidence } from "./createCompanyBodyConfidence";
import type { CreateCompanyBodyGeography } from "./createCompanyBodyGeography";
import type { CreateCompanyBodyTier } from "./createCompanyBodyTier";

export interface CreateCompanyBody {
  name: string;
  fullName: string;
  tier: CreateCompanyBodyTier;
  geography: CreateCompanyBodyGeography;
  exchange?: string;
  irPage?: string;
  earningsCadence?: string;
  revenue?: string;
  revenueGrowth?: string;
  ebitdaMargin?: string;
  marketCap?: string;
  categories?: string;
  strategicPriorities?: string;
  scIntelligence?: string;
  techIntelligence?: string;
  openProblems?: string;
  quickTake?: string;
  confidence?: CreateCompanyBodyConfidence;
  version?: string;
  nextRefresh?: string;
}