# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Industry Research Pod API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalCompanies: number;
  tier1Companies: number;
  tier2Companies: number;
  totalSheiCards: number;
  immediateSheiCards: number;
  totalSignals: number;
  highSignals: number;
  totalBenchmarks: number;
  totalPlaybookSections: number;
}